Add exception handling for WineController

+++ src/main/java/se/iths/corkdork/controller/WineController.java
@@ -1,9 +1,12 @@
 package se.iths.corkdork.controller;
 
+import org.jetbrains.annotations.NotNull;
 import se.iths.corkdork.entity.WineEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import se.iths.corkdork.exception.BadRequestException;
+import se.iths.corkdork.exception.EntityNotFoundException;
 import se.iths.corkdork.service.WineService;
 
 import java.util.Optional;
@@ -22,14 +25,16 @@ public class WineController {
 
     @PostMapping
     public ResponseEntity<WineEntity> createWine(@RequestBody WineEntity wineEntity){
-        WineEntity createdWine = wineService.createWine(wineEntity);
+        if(wineEntity.getName().isEmpty())
+            throw new BadRequestException("Name cannot be empty.");
 
+        WineEntity createdWine = wineService.createWine(wineEntity);
         return new ResponseEntity<>(createdWine, HttpStatus.CREATED);
     }
 
     @PatchMapping("{id}")
     public ResponseEntity<WineEntity> updateWineName(@PathVariable Long id, @RequestBody String name) {
-        return new ResponseEntity<>(wineService.updateWine(id, name), OK);
+        return new ResponseEntity<>(wineService.updateWineName(id, name), OK);
     }
 
     @PutMapping("{id}")
@@ -43,20 +48,33 @@ public class WineController {
 
     @DeleteMapping({"{id}"})
     public ResponseEntity<Void> deleteWine(@PathVariable Long id){
+        if(wineService.findWineById(id).isEmpty())
+            throw new EntityNotFoundException(notFound(id));
+
         wineService.deleteWine(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<WineEntity>> findWineById(@PathVariable Long id) {
-        Optional<WineEntity> foundWine = wineService.findWineById(id);
+        if(wineService.findWineById(id).isEmpty())
+            throw new EntityNotFoundException(notFound(id));
 
+        Optional<WineEntity> foundWine = wineService.findWineById(id);
         return new ResponseEntity<>(foundWine, OK);
     }
 
     @GetMapping
     public ResponseEntity<Iterable<WineEntity>> findAllWines(){
         Iterable<WineEntity> allWines = wineService.findAllWines();
+        if(!allWines.iterator().hasNext())
+            throw new EntityNotFoundException("Failed to find any wines.");
+
         return new ResponseEntity<>(allWines, OK);
+    }
+
+    @NotNull
+    private String notFound(Long id) {
+        return "Wine with ID: " + id + "was not found.";
     }
 }
